Allow boat discount based on number of units owned (#1261)
The boat discount cannot be abused and when you’re stuck on a small
island, kind of isolated and low on resources, being able to afford a
boat with your last few gold can. A last move resistance and fun to
play.

- [x] I have added screenshots for all UI updates
- [x] I process any text displayed to the user through translateText()
and I've added it to the en.json file
- [x] I have added relevant tests to the test directory
- [x] I confirm I have thoroughly tested these changes and take full
responsibility for any bugs introduced
- [x] I understand that submitting code with bugs that could have been
caught through manual testing blocks releases and new features for all
contributors
regression is found:

devalnor